+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Change Log
 - Use AnnotatedString for login prompts in Subjects
   - Show login prompt instead of interest buttons in subject detail for guests
 - Add back button to Login app bar
+- Fix group searches
 
 ## Version 0.7.6(706)